feat: middleware-related stuff separated from client
The original client.py was divided in two files,
client_base.py, where only NetApp communication is,
and client.py, which inherits from client_base and adds
the middleware-related communication.

The gstreamer part was integrated directly into the
client_base -> in the future, it would be nice to
make gstreamer and websocket communication as a plugin.

Examples adapted to the new client interface and made
more general.